[chore #158894615]refactor code
