+++ models/db.py
@@ -3,18 +3,11 @@ import os
 
 class Database:
     def __init__(self):
-        # self.host = "",
-        # self.database = "Ride_with_me",
-        # self.user = "naibor",
-        # self.password = "lisanaibor",
         self.connection = psycopg2.connect(
             host="localhost",
             user=os.getenv("DATABASE_USER","naibor"),
-            # user="naibor",
             dbname=os.getenv("DATABASE_NAME","Ride_with_me"),
-            # dbname="Ride_with_me",
-            password=os.getenv("DATABASE_PASSWORD","lisanaibor"),
-            # password="lisanaibor"
+            password=os.getenv("DATABASE_PASSWORD","lisanaibor")
         )
         self.cursor = self.connection.cursor()
 
@@ -24,13 +17,13 @@ class Database:
     def query_db(self, query_string):
        self.cursor.execute(query_string)
 
-    
     def create_tables(self, table):
         self.cursor.execute(table)
         self.commit()
 
     def drop_table(self,table_name):
         self.cursor.execute("DROP TABLE IF EXISTS " + table_name + ";") 
+        self.commit()
 
     def close(self):
         self.cursor.close()
